+++ src/main/java/domain/CarRace.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -23,7 +25,7 @@ public class CarRace {
     return updateAllWinners(findWinningDistance());
   }
 
-  public List<String> getCoWinners() {
+  public List<String> getWinners() {
     return coWinners;
   }
