+++ src/main/java/domain/RacingCar.java
@@ -1,3 +1,5 @@
+package domain;
+
 public class RacingCar {
   private final String name;
   private int distance;
